% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kelpbioshiny-package.R
\docType{package}
\name{kelpbioshiny-package}
\alias{kelpbioshiny}
\alias{kelpbioshiny-package}
\title{kelpbioshiny: Shiny App for Bayesian Kelp Biomass Estimation}
\description{
A 'shiny' app for annual estimates of kelp biomass by site with the 'kelpbio' package. It steps through loading survey data, fitting or choosing pre-fit sub-models, checking their diagnostics, and exporting the biomass estimates.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/HakaiInstitute/kelpbioshiny}
  \item Report bugs at \url{https://github.com/HakaiInstitute/kelpbioshiny/issues}
}

}
\author{
\strong{Maintainer}: Seb Dalgarno \email{seb@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-3658-4517}{ORCID})

Authors:
\itemize{
  \item Seb Dalgarno \email{seb@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-3658-4517}{ORCID})
}

Other contributors:
\itemize{
  \item Tula Foundation [funder, copyright holder]
}

}
\keyword{internal}
